requirements.txt
streamlit
pandas
numpy
scikit-learn
streamlit-chat 
openai
python-dotenv